created full module for trainer profile and cleaned user module
