+++ src/modules/users/users.controller.ts
@@ -6,19 +6,12 @@ import {
   HttpCode,
   HttpStatus,
   Param,
-  ParseIntPipe,
   Post,
   Put,
-  UseGuards,
 } from '@nestjs/common';
-import { User } from 'src/decorators/user.decorator';
 import { AddUserDto } from 'src/dto/AddUser.dto';
-import { ClientDto, CreateClientDto, GhostClientDto } from 'src/dto/client.dto';
-import { TrainerDto, TrainerProfileDto } from 'src/dto/trainer.dto';
-import { UserDto } from 'src/dto/user.dto';
+import { CreateClientDto, GhostClientDto } from 'src/dto/client.dto';
 import { VerificationDto } from 'src/dto/verification.dto';
-import { Trainer } from 'src/entites/trainer.entity';
-import { JwtAuthGuard } from '../auth/local-auth.guard';
 import { UsersService } from './users.service';
 
 @Controller('users')
@@ -50,22 +43,14 @@ export class UsersController {
     return this.usersService.remove(+id);
   }
 
-  @Post('assign-client-to-trainer/:clientId/:trainerId')
-  async assignClientToTrainer(
-    @Param('clientId') clientId: number,
-    @Param('trainerId') trainerId: number,
-  ): Promise<TrainerDto> {
-    return this.usersService.assignClientToTrainer(clientId, trainerId);
-  }
-
-  @Get(':userId/clients')
-  async getTrainerWithClients(
-    @Param('userId', ParseIntPipe) userId: number,
-  ): Promise<ClientDto[]> {
-    const trainerId = (await this.usersService.findTrainerIdByUserId(userId))
-      .id;
-    return this.usersService.findTrainerClients(trainerId);
-  }
+  // only trainer will assign clients to them
+  // to be fixed
+  // @Post('assign-client-to-trainer')
+  // async assignClientToTrainer(
+  //   @Body() clientTrainerObject:any
+  // ): Promise<TrainerDto> {
+  //   return this.usersService.assignClientToTrainer(clientTrainerObject);
+  // }
 
   @Post('verify')
   @HttpCode(HttpStatus.OK)
@@ -85,37 +70,10 @@ export class UsersController {
     return ghostClient;
   }
 
-  @Get('get-id-by-user/:userId')
-  async getTrainerIdByUser(
-    @Param('userId') userId: number,
-  ): Promise<number | null> {
-    return this.usersService.getTrainerIdByUser(userId);
-  }
-
-  @Post('trainer-profile')
-  @UseGuards(JwtAuthGuard)
-  updateTrainerProfile(
-    @User() user: UserDto,
-    @Body() trainerProfileDto: TrainerProfileDto,
-  ): Promise<Trainer | null> {
-    const userId = user.id;
-    return this.usersService.updateTrainerProfile(userId, trainerProfileDto);
-  }
-
-  @Get('trainer-profile')
-  @UseGuards(JwtAuthGuard)
-  async getTrainerProfile(@User() user:UserDto): Promise<Trainer> {
-    try {
-      const id = user.id;
-      const trainerProfile = await this.usersService.findTrainerIdByUserId(id);
-
-      if (!trainerProfile) {
-        throw new Error('Trainer profile not found');
-      }
-      return trainerProfile;
-    } catch (error) {
-      console.error(`Error fetching trainer profile: ${error.message}`);
-      throw new Error('Unable to fetch trainer profile');
-    }
-  }
+  // @Get('get-id-by-user/:userId')
+  // async getTrainerIdByUser(
+  //   @Param('userId') userId: number,
+  // ): Promise<number | null> {
+  //   return this.usersService.getTrainerIdByUser(userId);
+  // }
 }
